move store to redux folder
