+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import './index.css';
 import { Provider } from 'react-redux';
-import store from './store';
+import store from './redux/store';
 import reportWebVitals from './reportWebVitals';
 import App from './App';
